Return to "id" for transfers, and add "id" to approvals
After further thought, transfers don't have a "subject", but they do have
an ID after all. Approvals also need a similar ID. These should be guaranteed
unique and are used primarily for de-duplication.

The "subject" on approvals is a different thing altogether, used for identifying
the parties and parameters involved in the scope of an approval.

Signed-off-by: Andrew Richardson <andrew.richardson@kaleido.io>